+++ src/full/Agda/Syntax/Internal/Names.hs
@@ -4,7 +4,6 @@ module Agda.Syntax.Internal.Names where
 
 import Data.HashMap.Strict (HashMap)
 import qualified Data.HashMap.Strict as HMap
-import Data.List.NonEmpty (NonEmpty(..))
 import Data.Map (Map)
 import Data.Set (Set)
 
@@ -18,6 +17,7 @@ import Agda.Syntax.Treeless
 import Agda.TypeChecking.Monad.Base
 import Agda.TypeChecking.CompiledClause
 
+import Agda.Utils.List1 (List1)
 import qualified Agda.Utils.Maybe.Strict as Strict
 import Agda.Utils.Singleton
 import Agda.Utils.Impossible
@@ -73,18 +73,22 @@ class NamesIn a where
 instance NamesIn a => NamesIn (Maybe a)
 instance NamesIn a => NamesIn (Strict.Maybe a)
 instance NamesIn a => NamesIn [a]
-instance NamesIn a => NamesIn (NonEmpty a)
+instance NamesIn a => NamesIn (List1 a)
 instance NamesIn a => NamesIn (Set a)
 instance NamesIn a => NamesIn (Map k a)
 
 -- Decorations
 instance NamesIn a => NamesIn (Arg a)
-instance NamesIn a => NamesIn (Dom a)
 instance NamesIn a => NamesIn (Named n a)
 instance NamesIn a => NamesIn (Abs a)
 instance NamesIn a => NamesIn (WithArity a)
 instance NamesIn a => NamesIn (Open a)
 instance NamesIn a => NamesIn (C.FieldAssignment' a)
+
+instance (NamesIn a, NamesIn b) => NamesIn (Dom' a b) where
+  namesAndMetasIn' sg (Dom _ _ _ t e) =
+    mappend (namesAndMetasIn' sg t) (namesAndMetasIn' sg e)
+
 
 -- Specific collections
 instance NamesIn a => NamesIn (Tele a)
@@ -263,6 +267,7 @@ instance NamesIn a => NamesIn (Builtin a) where
   namesAndMetasIn' sg = \case
     Builtin t -> namesAndMetasIn' sg t
     Prim x    -> namesAndMetasIn' sg x
+    BuiltinRewriteRelations xs -> namesAndMetasIn' sg xs
 
 -- | Note that the 'primFunImplementation' is skipped.
 instance NamesIn PrimFun where
@@ -291,6 +296,7 @@ instance NamesIn NLPSort where
   namesAndMetasIn' sg = \case
     PType a       -> namesAndMetasIn' sg a
     PProp a       -> namesAndMetasIn' sg a
+    PSSet a       -> namesAndMetasIn' sg a
     PInf _ _      -> mempty
     PSizeUniv     -> mempty
     PLockUniv     -> mempty