+++ src/full/Agda/TypeChecking/Coverage/Cubical.hs
@@ -885,11 +885,12 @@ createMissingTrXConClause q_trX f n x old_sc c (UE gamma gamma' xTel u v rho tau
     ,  nest 2 $ prettyTCM . QNamed f $ cl
     ]
 
-  let mod = fromMaybe __IMPOSSIBLE__ $ scTarget old_sc
+  let mod =
+        setRelevance Irrelevant $  -- See #5611.
+        getModality $ fromMaybe __IMPOSSIBLE__ $ scTarget old_sc
   -- we follow what `cover` does when updating the modality from the target.
   applyModalityToContext mod $ do
     unlessM (asksTC hasQuantity0) $ do
-    let mod = getModality . fromJust $ scTarget old_sc
     reportSDoc "tc.cover.trxcon" 20 $ text "testing usable at mod: " <+> pretty mod
     addContext cTel $ usableAtModality mod rhs