+++ src/full/Agda/TypeChecking/Rewriting.hs
@@ -48,6 +48,7 @@ import Prelude hiding (null)
 
 import Control.Monad
 
+import Data.Foldable (toList)
 import Data.IntSet (IntSet)
 import qualified Data.IntSet as IntSet
 import qualified Data.List as List
@@ -160,6 +161,19 @@ addRewriteRules qs = do
     reportSDoc "rewriting" 10 $
       "done checking confluence of rules" <+> prettyList_ (map (prettyTCM . rewName) rews)
 
+-- Auxiliary function for checkRewriteRule.
+-- | Get domain of rewrite relation.
+rewriteRelationDom :: QName -> TCM (ListTel, Dom Type)
+rewriteRelationDom rel = do
+  -- We know that the type of rel is that of a relation.
+  relV <- relView =<< do defType <$> getConstInfo rel
+  let RelView _tel delta a _a' _core = fromMaybe __IMPOSSIBLE__ relV
+  reportSDoc "rewriting" 30 $ do
+    "rewrite relation at type " <+> do
+      inTopContext $ prettyTCM (telFromList delta) <+> " |- " <+> do
+        addContext delta $ prettyTCM a
+  return (delta, a)
+
 -- | Check the validity of @q : Γ → rel us lhs rhs@ as rewrite rule
 --   @
 --       Γ ⊢ lhs ↦ rhs : B
@@ -173,7 +187,11 @@ checkRewriteRule q = do
   requireOptionRewriting
   let failNoBuiltin = typeError $ GenericError $
         "Cannot add rewrite rule without prior BUILTIN REWRITE"
-  rel <- fromMaybeM failNoBuiltin $ getBuiltinName builtinRewrite
+  rels <- fromMaybeM failNoBuiltin getBuiltinRewriteRelations
+  reportSDoc "rewriting.relations" 40 $ vcat
+    [ "Rewrite relations:"
+    , prettyList $ map prettyTCM $ toList rels
+    ]
   def <- instantiateDef =<< getConstInfo q
   -- Issue 1651: Check that we are not adding a rewrite rule
   -- for a type signature whose body has not been type-checked yet.
@@ -183,13 +201,6 @@ checkRewriteRule q = do
       , prettyTCM q
       , " cannot be added before the function definition"
       ]
-  -- We know that the type of rel is that of a relation.
-  relV <- relView =<< do defType <$> getConstInfo rel
-  let RelView _tel delta a _a' _core = fromMaybe __IMPOSSIBLE__ relV
-  reportSDoc "rewriting" 30 $ do
-    "rewrite relation at type " <+> do
-      inTopContext $ prettyTCM (telFromList delta) <+> " |- " <+> do
-        addContext delta $ prettyTCM a
   -- Get rewrite rule (type of q).
   TelV gamma1 core <- telView $ defType def
   reportSDoc "rewriting" 30 $ vcat
@@ -209,13 +220,20 @@ checkRewriteRule q = do
   let failureFreeVars :: IntSet -> TCM a
       failureFreeVars xs = typeError . GenericDocError =<< hsep
         [ prettyTCM q , " is not a legal rewrite rule, since the following variables are not bound by the left hand side: " , prettyList_ (map (prettyTCM . var) $ IntSet.toList xs) ]
+  let failureNonLinearPars :: IntSet -> TCM a
+      failureNonLinearPars xs = typeError . GenericDocError =<< do
+        (prettyTCM q
+          <+> " is not a legal rewrite rule, since the following parameters are bound more than once on the left hand side: "
+          <+> hsep (List.intersperse "," $ map (prettyTCM . var) $ IntSet.toList xs))
+          <> ". Perhaps you can use a postulate instead of a constructor as the head symbol?"
   let failureIllegalRule :: TCM a -- TODO:: Defined but not used
       failureIllegalRule = typeError . GenericDocError =<< hsep
         [ prettyTCM q , " is not a legal rewrite rule" ]
 
   -- Check that type of q targets rel.
   case unEl core of
-    Def rel' es@(_:_:_) | rel == rel' -> do
+    Def rel es@(_:_:_) | rel `elem` rels -> do
+      (delta, a) <- rewriteRelationDom rel
       -- Because of the type of rel (Γ → sort), all es are applications.
       let vs = map unArg $ fromMaybe __IMPOSSIBLE__ $ allApplyElims es
       -- The last two arguments are lhs and rhs.
@@ -294,6 +312,10 @@ checkRewriteRule q = do
           "variables used by the rewrite rule: " <+> text (show usedVars)
         unlessNull (freeVars IntSet.\\ boundVars) failureFreeVars
         unlessNull (usedVars IntSet.\\ (boundVars `IntSet.union` IntSet.fromList pars)) failureFreeVars
+
+        reportSDoc "rewriting" 70 $
+          "variables bound in (erased) parameter position: " <+> text (show pars)
+        unlessNull (boundVars `IntSet.intersection` IntSet.fromList pars) failureNonLinearPars
 
         let rew = RewriteRule q gamma f ps rhs (unDom b) False