+++ src/full/Agda/TypeChecking/Rewriting/NonLinMatch.hs
@@ -224,6 +224,7 @@ instance Match () NLPSort Sort where
     case (p , s) of
       (PType lp  , Type l  ) -> match r gamma k () lp l
       (PProp lp  , Prop l  ) -> match r gamma k () lp l
+      (PSSet lp  , SSet l  ) -> match r gamma k () lp l
       (PInf fp np , Inf f n)
         | fp == f, np == n   -> yes
       (PSizeUniv , SizeUniv) -> yes