+++ src/full/Agda/TypeChecking/Serialise/Instances/Common.hs
@@ -13,7 +13,6 @@ import Data.Array.IArray
 import Data.Word
 import qualified Data.Foldable as Fold
 import Data.Hashable
-import qualified Data.HashTable.IO as H
 import Data.Int (Int32)
 
 import Data.Map (Map)
@@ -21,8 +20,6 @@ import qualified Data.Map as Map
 import Data.Set (Set)
 import qualified Data.IntSet as IntSet
 import Data.IntSet (IntSet)
-import Data.List.NonEmpty (NonEmpty(..), nonEmpty)
-import qualified Data.List.NonEmpty as NonEmpty
 import qualified Data.Set as Set
 import Data.Sequence (Seq)
 import qualified Data.Sequence as Seq
@@ -49,6 +46,9 @@ import Agda.Utils.BiMap (BiMap)
 import qualified Agda.Utils.BiMap as BiMap
 import qualified Agda.Utils.Empty as Empty
 import Agda.Utils.FileName
+import qualified Agda.Utils.HashTable as H
+import Agda.Utils.List1 (List1)
+import qualified Agda.Utils.List1 as List1
 import Agda.Utils.List2 (List2(List2))
 import qualified Agda.Utils.List2 as List2
 import Agda.Utils.Maybe
@@ -253,9 +253,9 @@ instance {-# OVERLAPPABLE #-} EmbPrj a => EmbPrj [a] where
 --                            valu [x, xs] = valu2 (:) x xs
 --                            valu _       = malformed
 
-instance EmbPrj a => EmbPrj (NonEmpty a) where
-  icod_ = icod_ . NonEmpty.toList
-  value = maybe malformed return . nonEmpty <=< value
+instance EmbPrj a => EmbPrj (List1 a) where
+  icod_ = icod_ . List1.toList
+  value = maybe malformed return . List1.nonEmpty <=< value
 
 instance EmbPrj a => EmbPrj (List2 a) where
   icod_ = icod_ . List2.toList
@@ -266,9 +266,31 @@ instance (EmbPrj k, EmbPrj v, EmbPrj (BiMap.Tag v)) =>
   icod_ m = icode (BiMap.toDistinctAscendingLists m)
   value m = BiMap.fromDistinctAscendingLists <$> value m
 
+
+-- | Encode a list of key-value pairs as a flat list.
+mapPairsIcode :: (EmbPrj k, EmbPrj v) => [(k, v)] -> S Int32
+mapPairsIcode xs = icodeNode =<< convert [] xs where
+  -- As we need to call `convert' in the tail position, the resulting list is
+  -- written (and read) in reverse order, with the highest pair first in the
+  -- resulting list.
+  convert ys [] = return ys
+  convert ys ((start, entry):xs) = do
+    start <- icode start
+    entry <- icode entry
+    convert (start:entry:ys) xs
+
+mapPairsValue :: (EmbPrj k, EmbPrj v) => [Int32] -> R [(k, v)]
+mapPairsValue = convert [] where
+  convert ys [] = return ys
+  convert ys (start:entry:xs) = do
+    start <- value start
+    entry <- value entry
+    convert ((start, entry):ys) xs
+  convert _ _ = malformed
+
 instance (Ord a, EmbPrj a, EmbPrj b) => EmbPrj (Map a b) where
-  icod_ m = icode (Map.toAscList m)
-  value m = Map.fromDistinctAscList <$> value m
+  icod_ m = mapPairsIcode (Map.toAscList m)
+  value = vcase (fmap Map.fromDistinctAscList . mapPairsValue)
 
 instance (Ord a, EmbPrj a) => EmbPrj (Set a) where
   icod_ s = icode (Set.toAscList s)
@@ -459,8 +481,8 @@ instance EmbPrj NameId where
   value = valueN NameId
 
 instance (Eq k, Hashable k, EmbPrj k, EmbPrj v) => EmbPrj (HashMap k v) where
-  icod_ m = icode (HMap.toList m)
-  value m = HMap.fromList `fmap` value m
+  icod_ m = mapPairsIcode (HMap.toList m)
+  value = vcase (fmap HMap.fromList . mapPairsValue)
 
 instance EmbPrj a => EmbPrj (WithHiding a) where
   icod_ (WithHiding a b) = icodeN' WithHiding a b